+++ R/summariseClinicalRecords.R
@@ -306,6 +306,7 @@ summariseRecordsPerPerson <- function(x, den, strata, estimates) {
     resultx |>
       dplyr::mutate(number_subjects = dplyr::if_else(.data$n == 0, 0L, 1L)) |>
       dplyr::select(!"person_id") |>
+      dplyr::collect() |>
       PatientProfiles::summariseResult(
         group = list(),
         includeOverallGroup = FALSE,
@@ -331,7 +332,7 @@ summariseRecordsPerPerson <- function(x, den, strata, estimates) {
       estimate_value = reduceDemicals(.data$estimate_value, 4)
     )
 
-  omopgenerics::dropTable(cdm = cdm, name = dplyr::starts_with(prefix))
+  omopgenerics::dropSourceTable(cdm = cdm, name = dplyr::starts_with(prefix))
 
   return(result)
 }